requirements.txt
numpy
pandas
scipy
numpy_financial
platypus-opt
streamlit
plotly
openpyxl

